% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances_clust.R,
%   R/distances_clust_start_endcol.R
\name{distances_clust}
\alias{distances_clust}
\title{distances_clust}
\usage{
distances_clust(morph, traits, weights, startCol = 7, endCol = NULL,
  contTraits = TRUE, coresFree = 2)

distances_clust(morph, traits, weights, startCol = 7, endCol = NULL,
  contTraits = TRUE, coresFree = 2)
}
\arguments{
\item{morph}{a data frame with (at least) columns for unique tooth identifiers
(IDCol specified below) and columns containing morphometric data to input into
the disparity calculation. Other columns can also be included in this data frame,
but will be ignord by the function, as specified below.}

\item{traits}{a list of trait connectivity matices. (For ichthyolith analyses based
after Sibert et al 2019, you can use use the 'traits' object included in this package,
which the function will include if you do not spcify a traits list.)}

\item{weights}{a vector of how much each trait is weighted in the analysis
can use data(weights) included in this package, or define your own.}

\item{startCol}{a numeric value referring to the column which contains the first
character trait in the morph matrix. Allows for differeing numbers of columns
of metadata for each object}

\item{endCol}{a numeric value referring to the column which contains the last
character trait in the morph matrix. If no value, this is assumed to be the last
column in the morph matrix.}

\item{coresFree}{a numerical value denoting how many cores to leave free
on your machine when running in parallel. Defaults to 2.
Use detectCores() to determine the number of cores available on the computer
to use. The number of cores in the cluster created in this function is
detectCores()-coresFree}

\item{morphCols}{a vetor stating the columns with morphometric data to pull for analysis.
This can be a continous list (e.g.c(1:9) to take columns 1-9 as morphometric data) or it
can be a subset (e.g. c(1:3, 5, 7:9) to exclude columns 6 and 8).}

\item{traitsStartCol}{is the numeric value of the column in the morph data frame
that matches to the first character in the traits list of connectivity matrices.}

\item{IDCol}{is a numeric value referring to the column in the morph data frame which
contains the unique identifiers for each object.}

\item{conTraits}{######### ideally a character vector containing the column names
of the continuous traits to consider, or a numeric vector containing the column numbers. ########}

\item{morph}{a data frame of the form returned by toothdat_cleanup}

\item{traits}{a list of trait connectivity matices (for standard datasets
use data(traits) included in this package}

\item{weights}{a vector of how much each trait is weighted in the analysis
can use data(weights) included in this package, or define your own.}

\item{conTraits}{######### ideally a character vector containing the column names
of the continuous traits to consider, or a numeric vector containing the column numbers. ########}

\item{coresFree}{a numerical value denoting how many cores to leave free
on your machine when running in parallel. Defaults to 2.
Use detectCores() to determine the number of cores available on the computer
to use. The number of cores in the cluster created in this function is
detectCores()-coresFree}
}
\value{
Returns a dataframe of each pairwise comparison with the
following columns: \cr
"1" and "2" - the combination of teeth being considered \cr
"ToothA" and "ToothB", - the row number for each tooth from the original input dataset (will be different from the 1 and 2 column only if objects were previously removed from the dataset) \cr
"dist.sum" - the summed distance between the two teeth \cr
"dist.avg" - the averaged distance between the two teeth \cr
"traits.length" - the number of traits actually considered in the comparison \cr
"objID.A" and "objID.B" - the unique identifiers for each tooth considered.

Returns a dataframe of each pairwise comparison with the
following columns: \cr
"1" and "2" - the combination of teeth being considered \cr
"ToothA" and "ToothB", - the row number for each tooth from the original input dataset \cr
"dist.sum" - the summed distance between the two teeth \cr
"dist.avg" - the averaged distance between the two teeth \cr
"traits.length" - the number of traits actually considered in the comparison \cr
"objID.A" and "objID.B" - the unique identifiers for each tooth considered.
}
\description{
The distances_clust function is the heart of the ichthyoliths package. It
calculates the morphological disparity of teeth by summing the distances
between each character/trait for each pair of teeth in the dataset.
It relies on cluster computing, and is dependent on the doParallel package.
To run this calculation without using parallelization, use the
\code{\link{distances_no_clust}} function

The distances_clust function is the heart of the ichthyoliths package. It
calculates the morphological disparity of teeth by summing the distances
between each character/trait for each pair of teeth in the dataset.
It relies on cluster computing, and is dependent on the doParallel package.
To run this calculation without using parallelization, use the
\code{\link{distances_no_clust}} function
}
