% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances_clust_old.R
\name{distances_clust_old}
\alias{distances_clust_old}
\title{distances_clust_old}
\usage{
distances_clust_old(morph, traits, weights, coresFree = 2)
}
\arguments{
\item{morph}{a data frame of the form returned by toothdat_cleanup}

\item{traits}{a list of trait connectivity matices (for standard datasets
use data(traits) included in this package}

\item{weights}{a vector of how much each trait is weighted in the analysis
can use data(weights) included in this package, or define your own.}

\item{coresFree}{a numerical value denoting how many cores to leave free
on your machine when running in parallel. Defaults to 2.
Use detectCores() to determine the number of cores available on the computer
to use. The number of cores in the cluster created in this function is
detectCores()-coresFree}
}
\value{
Returns a dataframe of each pairwise comparison with the
following columns: \cr
"1" and "2" - the combination of teeth being considered \cr
"ToothA" and "ToothB", - the row number for each tooth from the original input dataset \cr
"dist.sum" - the summed distance between the two teeth \cr
"dist.avg" - the averaged distance between the two teeth \cr
"traits.length" - the number of traits actually considered in the comparison \cr
"objID.A" and "objID.B" - the unique identifiers for each tooth considered.
}
\description{
The distances_clust_old function is the heart of the ichthyoliths package. It
calculates the morphological disparity of teeth by summing the distances
between each character/trait for each pair of teeth in the dataset.
It relies on cluster computing, and is dependent on the doParallel package.
To run this calculation without using parallelization, use the
\code{\link{distances_no_clust}} function
}
