% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances_clust_start_endcol.R
\name{distances_clust_start_endcol}
\alias{distances_clust_start_endcol}
\title{distances_clust_start_endcol}
\usage{
distances_clust_start_endcol(morph, traits, weights, startCol = 7,
  endCol = NULL, contTraits = TRUE, coresFree = 2)
}
\arguments{
\item{morph}{a data frame of the form returned by toothdat_cleanup}

\item{traits}{a list of trait connectivity matices (for standard datasets
use data(traits) included in this package}

\item{weights}{a vector of how much each trait is weighted in the analysis
can use data(weights) included in this package, or define your own.}

\item{startCol}{a numeric value referring to the column which contains the first
character trait in the morph matrix. Allows for differeing numbers of columns
of metadata for each object}

\item{endCol}{a numeric value referring to the column which contains the last
character trait in the morph matrix. If no value, this is assumed to be the last
column in the morph matrix.}

\item{coresFree}{a numerical value denoting how many cores to leave free
on your machine when running in parallel. Defaults to 2.
Use detectCores() to determine the number of cores available on the computer
to use. The number of cores in the cluster created in this function is
detectCores()-coresFree}

\item{conTraits}{######### ideally a character vector containing the column names
of the continuous traits to consider, or a numeric vector containing the column numbers. ########}
}
\value{
Returns a dataframe of each pairwise comparison with the
following columns: \cr
"1" and "2" - the combination of teeth being considered \cr
"ToothA" and "ToothB", - the row number for each tooth from the original input dataset \cr
"dist.sum" - the summed distance between the two teeth \cr
"dist.avg" - the averaged distance between the two teeth \cr
"traits.length" - the number of traits actually considered in the comparison \cr
"objID.A" and "objID.B" - the unique identifiers for each tooth considered.
}
\description{
The distances_clust function is the heart of the ichthyoliths package. It
calculates the morphological disparity of teeth by summing the distances
between each character/trait for each pair of teeth in the dataset.
It relies on cluster computing, and is dependent on the doParallel package.
To run this calculation without using parallelization, use the
\code{\link{distances_no_clust}} function
}
