% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances_no_clust.R
\name{distances_no_clust}
\alias{distances_no_clust}
\title{distances_no_clust}
\usage{
distances_no_clust(morph, traits, weights)
}
\arguments{
\item{morph}{a data frame of the form returned by toothdat_cleanup}

\item{traits}{a list of trait connectivity matices (for standard datasets
use data(traits) included in this package}

\item{weights}{a vector of how much each trait is weighted in the analysis
can use data(weights) included in this package, or define your own.}
}
\value{
Returns a dataframe of each pairwise comparison with the
following columns: \cr
"1" and "2" - the combination of teeth being considered \cr
"ToothA" and "ToothB", - the row number for each tooth from the original input dataset \cr
"dist.sum" - the summed distance between the two teeth \cr
"dist.avg" - the averaged distance between the two teeth \cr
"traits.length" - the number of traits actually considered in the comparison \cr
"objID.A" and "objID.B" - the unique identifiers for each tooth considered.
}
\description{
The distances_no_clust function is a variation on the distances_clust function,
the heart of the ichthyoliths package. It calculates the morphological disparity
of teeth by summing the distances between each character/trait for each pair
of teeth in the dataset. This function relies on the foreach loop, but does
\strong{NOT} create a cluster or run the loop in parallel, unlike
\code{\link{distances_clust}}.
}
