% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distmat.R
\name{distmat}
\alias{distmat}
\title{distmat}
\usage{
distmat(distpairs, type = "avg")
}
\arguments{
\item{distpairs}{an output data frame from distances_clust or distances_no_clust}

\item{type}{a string either c("avg", "sum") tells the function which calculated
distance value to populate the matrix with.}
}
\value{
distance matrix of dimensions nxn where n is the number of teeth considered.
}
\description{
The distmat function takes output from the \code{\link{distances_clust}} or
\code{\link{distances_no_clust}} function and coerces it into a distance matrix.
}
