% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_traits_csvs.R
\name{import_traits_csvs}
\alias{import_traits_csvs}
\title{import_traits_csvs}
\usage{
import_traits_csvs(csvname = "Trait*", csvpath = ".", recurs = T,
  ...)
}
\arguments{
\item{csvname}{a regular expression, Wildcard search term to find all
connectivity matrix csv files. Defaults to "Trait*". Case sensitive.}

\item{csvpath}{a character vector of full path names. Directory to look for
the trait connectivity csv files. Default correxponds to the  working
directory, but you can specify a relative or absolute path here.}

\item{recurs}{a logical value. Determines whether the search should recurse
into the directory. Defaults to TRUE.}

\item{...}{arguments to be passed to the list.files() function.}
}
\value{
A list of normalized connectivity matrices. Names of objects in the
list are the csv names without the .csv part
}
\description{
Imports connectivity matrices and builds list of all traits to include in
analysis. Use this if you are using connectivity matrices that are \strong{not}
described in the current version of ichthyolith classification for the package.
If using the standard ichthyolith coding scheme, simply call the included
traits data: \code{data(traits)}
}
\examples{
trait.matrices <- import_traits_csvs(csvname = "Trait*", csvpath = ".", recurs = T)

}
