% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangechart.R
\name{rangechart}
\alias{rangechart}
\title{rangechart}
\usage{
rangechart(
  counts,
  ages = NULL,
  taxa = NULL,
  tax.cat = NULL,
  reorder = NULL,
  normalize.counts = FALSE,
  count.breaks = NULL,
  cex.xaxis = 1,
  cex.yaxis = 1,
  yaxis.ticks = FALSE,
  llwd = 1,
  llcol = "gray70",
  llty = 3,
  baselines = FALSE,
  blwd = 0.5,
  blcol = "lightblue",
  blty = 3,
  rangeExtensions = NULL,
  elwd = 0.8,
  elcol = "red",
  elty = 2,
  col.points = "gray70",
  cols.vec = NULL,
  pch.points = 16,
  pch.vec = NULL,
  cex.points = 1,
  largesize = 1,
  xaxis.labels = c("names", "numeric", "alphanum"),
  print.xaxis = FALSE,
  xlab = "",
  ylab = "Age (Ma)",
  ...
)
}
\arguments{
\item{counts}{is the table, matrix, or data.frame of occurrances for each taxa (columns)
by age (rows). NOTE: Rows and Columns MUST have names for this to work, unless ages is specified.}

\item{ages}{is a vector giving the age for each sample depth. Default is to pull rownames of counts}

\item{taxa}{is a vector giving the name for each taxa. Default is to pull column-names of counts}

\item{tax.cat}{is an optional numerical vector that splits the taxa into categories or groups.
Must be equal to the length of the taxa vector. Must be numeric. Default is NULL}

\item{reorder}{is an optional command that allows the reordering of taxa to show extinction or
origination. It currently has two options:

a) 'fad.by.lad': Sort first by First Appearence Datum (FAD) then by Last Appearence Datum (LAD);
Highlights orrigination

b) 'lad.by.fad': Sort first by Last Appearence Datum (LAD) then by First Appearence Datum (FAD)
Highlights extinctions.}

\item{normalize.counts}{is a logical. When TRUE, the counts matrix is recalculated to
represent each taxa's occurrance counts as a percentage for each sample (age).
When FALSE, the raw count values are used. Default is FALSE.}

\item{count.breaks}{is a numerical input vector which gives the breaks between grouped
count values, as whole integer numbers. These can be percentages if normalize.counts=TRUE
or count values if normalize.counts = FALSE. If left blank, defaults to NULL. Note that
vector should take the form c(0, breakpoints). An example of this is
count.breaks = c(0, 2, 5, 10) will yield groups of <2%, 2-5%, 5-10%, and 10+%, or
raw counts of 1-2, 3-5, 6-10, and 10+.}

\item{cex.xaxis}{is a single numerical value scaling the x-axis labels}

\item{cex.yaxis}{is a single numerical value scaling the y-axis labels}

\item{yaxis.ticks}{is a logical. When TRUE, small ticks are added to the y-axis for
each sample's precise age. When FALSE, these ticks are not included.}

\item{llwd}{is the lwd (line width) parameter passed to the line segments connecting
the first and last occurrances points for each taxa.}

\item{llcol}{is the col (line color) parameter passed to the line segments connecting
the first and last  occurrances points for each taxa.}

\item{llty}{is the lty (line type) parameter passed to the line segments connecting
the first and last  occurrances points for each taxa.}

\item{baselines}{is a logical determining whether to draw line connecting the first
occurrance point to the x-axis. Default is FALSE (no baselines).}

\item{blwd}{is the lwd (line width) parameter passed to the baselines segment function.}

\item{blcol}{is the col (line color) parameter passed to the baselines segment function.}

\item{blty}{is the lty (line type) parameter passed to the baselines segment function.}

\item{rangeExtensions}{takes a vector of calculated range extension LADs and adds them to
the rangechart plot. Note that this numerical vector must be in the same order and have the same
length as the counts table that is fed to the function, and takes the CALCULATED range
extension LAD value. the rangechart function does not calculate range extensions for you.}

\item{elwd}{is the range extension line width}

\item{elcol}{is the range extension line color}

\item{elty}{is the range extension line type}

\item{col.points}{tells the function what color scheme to use:

a) specify a single color for all of the points. Do not specify a cols.vec

b) 'by.count' - the color will change based on the number or percentage of objects
in each age bin. Must specify a cols.vec.

c) 'by.category' - the color will change based on the tax.cat vector. Must specify a cols.vec}

\item{cols.vec}{is the color vector used to color the points. This can be internally defined
using a pre-existing color pallette (currently can use 'rainbow' from base or 'viridis' if
you have that package installed). If NULL and col.points is unspecified, the default color
of the points is gray70.

This can also be a user-defined vector of colors, but must be the length of the maximum
count value (if no count.breaks specified) or the number of groups defined by count.breaks
or tax.cat, if 'by.category' is selected in col.points. If insufficient colors are specified,
not all points will be plotted on the figure}

\item{pch.points}{tells the function what point symbol to use:

a) can be a single value defining the symbol for all points. Do not specify a pch.vec

b) 'by.count' - the point symbol will change based on the number or percentage of objects.
Must specify a pch.vec

c) 'by.category' - the point symbol will change based on the tax.cat vector. Must speciy a pch.vec}

\item{pch.vec}{is a vector used to define point shape. Must be the length of the number
of groups defined by count.breaks or tax.cat depending on what is selected for pch.points.}

\item{cex.points}{tells the function what point size to use:

a) can be a single value for point size, default is 1

b) 'by.count' - the size of the points will vary based on the number or percentage of objects.
Scales with count values.

c) 'by.category' - the size of points will be related to the category number specified in tax.cat}

\item{largesize}{is a scaling value to increase or decrease the size of the points.
Useful for scaling up large range charts. Default is 1.}

\item{xaxis.labels}{defines the values that will be put on the x-axis to denote taxa.

a) 'names' means that the names of the types will be the x-axis labels

b) 'numeric' means that they will be numbered 1-n on the x-axis.

c) 'alphanum' means that the original taxa order is preserved, and the numbers on
the x-axis refer to the original order of taxa in the counts matrix.}

\item{print.xaxis}{is a logical. If print.axis = TRUE, the function will return a list
of the corresponding morphotype names for each x-axis numeric value in the console.}
}
\value{
Returns a range chart style plot summarizing the counts table
}
\description{
The rangechart function takes an input table or matrix of ages (rows) and taxa (columns),
populated by occurrance or count values. This function does not create a legend, but as
it creates an active graphics device, you can add one based on the parameters defined here.
}
