% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toothdat_cleanup.R
\name{toothdat_cleanup}
\alias{toothdat_cleanup}
\title{toothdat_cleanup}
\usage{
toothdat_cleanup(toothdat, fix_dat = FALSE, sortby = "age-obj",
  toothdims = TRUE, mindim = 100)
}
\arguments{
\item{toothdat}{a data frame of teeth including columns for identification
of the teeth, as well as character/state values for each tooth and trait.
This function is most useful for use if the input csv file has a precise
format, detailed below.}

\item{fix_dat}{a logical value, defaults to false. If true, the values of
trait K1 and trait O will have 1 added to them. Use fix_dat=T if, when
entering data, you considered absence of a flange (trait K1) or absence of
a pulp cavity (trait O)  to have value of 0 (intuitive), rather than 1
(as defined in the paper).}

\item{sortby}{a string with value "age", "morph", "original", "age-obj".
Tells the function what order you would like the cleaned data frame to return
the individual tooth objects in. Assumes dataframe includes columns 1:4 as
detailed below. If "original", returns the data frame in its original order}

\item{toothdims}{a logical value that indicates whether the input dataframe
includes length, width, and aspect ratio measurements. Note that if this
value is true, in the case where some teeth have measurements and others do
not, a dummy value of AR=99 should be used to denote any teeth which have
no length/width measurements (e.g. are preserved as fragmens so measurements
do not have any significance), but which have been described using the
morphological characters.}

\item{mindim}{used only if toothdims = TRUE; a numerical value indicating
the minimum size of teeth to be considered in the dataset. Acts as a filter
to remove small objects.}
}
\value{
a reduced dataframe of just the tooth objects which have sufficient
morphological description to be analyzed. Includes all the original columns
}
\description{
Provides some code to clean up and sort an input data frame containing
individual tooth morphology into a dataframe usable by the distances_clust
function. Works best if input dataframe has the following columns:
Column 1: sample id - this can be anything and is not used by the function
Column 2: a numeric value (usually age of the sample) - used for sortby = 'age'
Column 3: an object number used for sortby = 'morph'.
  note that sortby = 'morph' should not be used for datasets that have
  multiple samples, as this number can be repeated
Column 4: an alphanumeric code correlating to the sample id (col 1) and an
  object identification (col 3). Used for sortby = "age-obj"
}
\examples{
toothdat <- toothdat_cleanup(toothdat, fix_dat=FALSE,
  sortby='age-obj', toothdims = TRUE, mindim = 100)

}
