% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_code.R
\name{validate_code}
\alias{validate_code}
\title{validate_code}
\usage{
validate_code(code.df, code.version = c("tooth_v0.5, denticle_v0.5"))
}
\arguments{
\item{code.df}{is the data frame which contains the morpholgoical character codes}

\item{code.version}{is a string specifying the version of validation to use.
Current options are c('tooth_v0.5', 'denticle_v0.5')}
}
\value{
Returns a data frame specifying:
   The character that is incorrectly coded ("character" column)
   The index of the tooth that is incorrectly coded ("index" column)
   The value that is currently coded ("incorrect.value" column)
This output can be used to then
}
\description{
The validate_code function checks whether the imported morphology code spreadsheet
contains correct values for each character. This is a useful check especially for
google sheets imports, which may have copy-paste errors that shuffle dropdowns
erroneously between columns.
}
