% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ichthyoliths.R
\docType{_PACKAGE}
\name{ichthyoliths}
\alias{ichthyoliths-package}
\alias{ichthyoliths}
\title{ichthyoliths}
\description{
A package for calculating morphological disparity of ichthyoliths (fish teeth and shark scales)
}
\details{
The ichthyolith package provides functions to define trait disparity matrices,
calculate tooth disparity, and create range charts and figures through time.

The disparity method used in this version of the ichthyoliths package (0.4, to
be renamed 1.0 upon publication) is described in detail in

The original version of the code (v0.1) is described in detail in
Sibert et al (2018), "Two pulses of morphological diversification in Pacific
pelagic fishes following the Cretaceous–Palaeogene mass extinction" (2018)
Proceedings of the Royal Society B: Biological Sciences
http://doi.org/10.1098/rspb.2018.1194.


For details and questions please contact Elizabeth Sibert \email{esibert@whoi.edu}.

At the heart of the ichthyoliths package are functions to calculate
tooth disparity ichthyoliths included in the analysis. They take advantage
of parallel computing, as pairwise-comparisons can grow quickly.This relies
on the package doParallel, and uses a foreach loop.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/esibert/ichthyoliths}
  \item Report bugs at \url{https://github.com/esibert/ichthyoliths/issues}
}

}
\author{
\strong{Maintainer}: Elizabeth Sibert \email{esibert@whoi.edu}

}
